problem 1/requirement.txt
django
psycopg
djangorestframework-simplejwt
djangorestframework
django-tenants